added support for jvm_flags and multiple custom suffixes
