+++ scala/scala.bzl
@@ -299,15 +299,16 @@ def _write_launcher(ctx, jars):
 
 
 def _write_junit_test_launcher(ctx, jars, test_suite):
-    #TODO- allow jvm flags for the junit process
     content = """#!/bin/bash
-{java} -cp {cp} -ea {classesFlag} org.junit.runner.JUnitCore {test_suite_class} 
+{java} -cp {cp} -ea {classesFlag} {jvm_flags} org.junit.runner.JUnitCore {test_suite_class} 
 """
     content = content.format(
       java=ctx.file._java.short_path,
       cp=":".join([j.short_path for j in jars]),
       test_suite_class=test_suite.suite_class,
-      classesFlag = test_suite.classesFlag
+      classesFlag = test_suite.classesFlag,
+      #allows setting xmx for example for tests which use a lot of memory
+      jvm_flags = " ".join(ctx.attr.jvm_flags)
     )
     ctx.file_action(
         output=ctx.outputs.executable,
@@ -509,26 +510,25 @@ def _scala_test_impl(ctx):
     _write_test_launcher(ctx, rjars)
     return _scala_binary_common(ctx, cjars, rjars)
 
-def _discover_classes(ctx, suffix, archive):
+def _discover_classes(ctx, suffixes, archive):
     discovered_classes = ctx.new_file(ctx.label.name + "_discovered_classes.txt")
     ctx.action(
       inputs=[archive],
       outputs=[discovered_classes],
-      progress_message="Discovering classes with suffix of %s" % suffix,
+      progress_message="Discovering classes with suffixes of %s" % suffixes,
       #TODO consider with Damien/Ulf/Oscar the implications of switching from grep to scala code
       #Pro-> logic will be cohesive (currently the scala code assumes stuff from the grep)
       #Con-> IIRC Ulf warned me about performance implications of these traversals
-      command="unzip -l {archive} | grep {suffix}.class > {out}".format(archive=archive.path, suffix=suffix,out=discovered_classes.path))
+      command="unzip -l {archive} | grep -e {combined_suffixes}.class > {out}".format(archive=archive.path,combined_suffixes=" -e ".join(suffixes),out=discovered_classes.path))
     return discovered_classes
 
 def _gen_test_suite_based_on_prefix(ctx, archive):
     #TODO enable custom patterns
-    #TODO evolve from suffix to pattern and add "Test*","*IT","*E2E" by default
-    discovered_classes = _discover_classes(ctx, "Test", archive)
+    #TODO evolve from suffixes to pattern and add "Test*","*IT","*E2E" by default
+    discovered_classes = _discover_classes(ctx, ctx.attr.suffixes, archive)
     return struct(suite_class = "io.bazel.rulesscala.test_discovery.DiscoveredTestSuite", classesFlag = "-Dbazel.discovered.classes.file.path=%s" % discovered_classes.short_path, discovered_classes = discovered_classes)
 
 def _scala_junit_test_impl(ctx):
-#TODO write a test that needs a compile time dependency and a runtime dependency    
     deps = ctx.attr.deps + [ctx.attr._suite]
     jars = _collect_jars(deps)
     (cjars, rjars) = (jars.compiletime, jars.runtime)
@@ -794,6 +794,7 @@ scala_junit_test = rule(
   attrs={
 #TODO see if we can parameterize the runner
 #      "main_class": attr.string(default="org.junit.runner.JUnitCore"),
+      "suffixes": attr.string_list(default=["Test"]),
       "_junit": attr.label(default=Label("//external:io_bazel_rules_scala/dependency/junit/junit"), single_file=True),
       "_hamcrest": attr.label(default=Label("//external:io_bazel_rules_scala/dependency/hamcrest/hamcrest_core"), single_file=True),
       "_suite": attr.label(default=Label("//src/java/io/bazel/rulesscala/test_discovery:test_discovery")),
